Adding casts to avoid warnings

+++ src/game/utils/utf8.cc
@@ -136,7 +136,7 @@ utf8_to_wchar(const char *in, size_t insize, wchar_t *out, size_t outsize,
 		}
 
 		/* does the sequence header tell us truth about length? */
-		if (lim - p <= n - 1) {
+		if (lim - p <= (signed int) n - 1) {
 			if ((flags & UTF8_IGNORE_ERROR) == 0)
 				return (0);
 			n = 1;
@@ -263,7 +263,7 @@ wchar_to_utf8(const wchar_t *in, size_t insize, char *out, size_t outsize,
 		if (out == NULL)
 			continue;
 
-		if (lim - p <= n - 1)
+		if (lim - p <= (signed int) n - 1)
 			return (0);		/* no space left */
 
 		/* make it work under different endians */
